* getndelim2.c: Make limit an ssize_t rather than int, as per the return type. * getndelim2.h: Ditto. * getnline.c: Ditto. * getnline.h: Ditto.
CVS-ID: lib/ChangeLog 1.206
CVS-ID: lib/getndelim2.c 1.3
CVS-ID: lib/getndelim2.h 1.3
CVS-ID: lib/getnline.c 1.5
CVS-ID: lib/getnline.h 1.5

+++ lib/getndelim2.c
@@ -52,7 +52,7 @@ char *malloc (), *realloc ();
    including the NUL terminator), or -1 on error or EOF.  */
 
 ssize_t
-getndelim2 (char **lineptr, size_t *linesize, size_t offset, int limit,
+getndelim2 (char **lineptr, size_t *linesize, size_t offset, ssize_t limit,
             int delim1, int delim2, FILE *stream)
 {
   size_t nbytes_avail;		/* Allocated but unused chars in *LINEPTR.  */
